todo-list/server/requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
email-validator
python-jose[cryptography]
bcrypt>=4.0.1
pytest
